clients/web: refactor TransactionsList
