+++ clients/apps/web/src/app/(main)/dashboard/[organization]/(header)/finance/(wide)/payouts/PayoutsPage.tsx
@@ -9,7 +9,6 @@ import DownloadInvoice, {
 import { PayoutProvider } from '@/components/Payouts/PayoutContext'
 import { PayoutStatus } from '@/components/Payouts/PayoutStatus'
 import AccountBanner from '@/components/Transactions/AccountBanner'
-import { platformFeesDisplayNames } from '@/components/Transactions/TransactionsList'
 import { useOrganizationAccount } from '@/hooks/queries'
 import { usePayouts } from '@/hooks/queries/payouts'
 import { getServerURL } from '@/utils/api'
@@ -19,6 +18,7 @@ import {
   getAPIParams,
   serializeSearchParams,
 } from '@/utils/datatable'
+import { platformFeesDisplayNames } from '@/utils/transaction'
 import { ClientResponseError, schemas } from '@polar-sh/client'
 import { formatCurrency } from '@polar-sh/currency'
 import Button from '@polar-sh/ui/components/atoms/Button'
